requirements.txt
# LLM API 客户端
anthropic>=0.25.0
openai>=1.0.0
httpx>=0.24.0  # DeepSeek API 使用

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 配置和环境
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 工具
typing-extensions>=4.7.0
dataclasses-json>=0.6.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# 可视化（可选）
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# JSON 操作
orjson>=3.9.0

